using subclasses for Edit and Add FeelDialogs
